updated folder structure naming
