+++ met-web/src/App.tsx
@@ -12,7 +12,7 @@ import AuthenticatedRoutes from './routes/AuthenticatedRoutes';
 import { Notification } from 'components/common/notification';
 import PageViewTracker from 'routes/PageViewTracker';
 import { NotificationModal } from 'components/common/modal';
-import { FeedbackModal } from 'components/common/Modals/Feedback/FeedbackModal';
+import { FeedbackModal } from 'components/common/Modals/Feedback';
 
 const App = () => {
     const drawerWidth = 240;
